+++ src/assets/CHANGELOG.md
@@ -1,11 +1,15 @@
 ## 2.2.0 (2018-12-22)
 
+#### New
++ **ngDraggable**: add [lockAxis] input to restrict dragging to a specific axis by locking another one.
+
 #### Bugfix
 + **ngDraggable**:
-    + Performance update. Fix [issue #112](https://github.com/xieziyu/angular2-draggable/issues/112): Control change detection with HostListener events.
-    + Fix [issue #128](https://github.com/xieziyu/angular2-draggable/issues/128): Multiple phone draggables at the same time.
+    + fix [issue #112](https://github.com/xieziyu/angular2-draggable/issues/112): Control change detection with HostListener events. Performance updated.
+    + fix [issue #128](https://github.com/xieziyu/angular2-draggable/issues/128): Multiple phone draggables at the same time.
 + **ngResizable**:
-    + Fix [issue #132](https://github.com/xieziyu/angular2-draggable/issues/132): Aspect ratio feature exits Y-Axis boundary on resize.
+    + fix [issue #132](https://github.com/xieziyu/angular2-draggable/issues/132): Aspect ratio feature exits Y-Axis boundary on resize.
+    + Performance updated.
 ---
 
 ## 2.1.9 (2018-11-29)